Added saving to file and reading from file.

+++ src/main/java/org/vut/ija_project/DataLayer/Robot/ControlledRobot.java
@@ -124,7 +124,7 @@ public class ControlledRobot implements Robot {
     }
 
     @Override
-    public double getRotationAngle() {
+    public int getRotationAngle() {
         return 0;
     }
